refactor(utils): refactor model utils
