+++ omnisafe/models/actor_critic.py
@@ -23,7 +23,7 @@ from gymnasium.spaces import Box, Discrete
 
 from omnisafe.models.actor import ActorBuilder
 from omnisafe.models.critic import CriticBuilder
-from omnisafe.utils.model_utils import build_mlp_network
+from omnisafe.utils.model import build_mlp_network
 
 
 # pylint: disable-next=too-many-instance-attributes
